+++ vss_cli/const.py
@@ -5,7 +5,7 @@ import pkg_resources
 
 PACKAGE_NAME = "vss_cli"
 
-__version__ = "2023.6.0"
+__version__ = "2023.6.1"
 
 
 DEFAULT_TIMEOUT = 30