+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog 📝
 
+## [v2023.6.1](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v2023.6.1) (2023-06-26)
+
+**Improvements:**
+
+- [#629](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/629): ``compute vm set ubuntu-pro``: attach/detach commands
+- [#630](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/630): ``core``: upgrade ``pyvss`` from ``2023.2.1`` to ``2023.6.0``.
+- [#631](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/631): ``core``: check motd via ``pyvss``.
+- [#632](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/632): ``domain``: update help.
+
 ## [v2023.6.0](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v2023.6.0) (2023-06-14)
 
 **Improvements:**